+++ src/libraries/System.Net.Http/tests/FunctionalTests/NtAuthTests.cs
@@ -104,7 +104,7 @@ namespace System.Net.Http.Functional.Tests
         }
     }
 
-    public class NtAuthTests : IClassFixture<NtAuthServers>
+    public partial class NtAuthTests : IClassFixture<NtAuthServers>
     {
         private readonly NtAuthServers _servers;
         private readonly ITestOutputHelper _output;
